move parseTrace
