+++ src/sentry/static/sentry/app/components/events/interfaces/spans/types.tsx
@@ -59,3 +59,11 @@ export enum TickAlignment {
   Right,
   Center,
 }
+
+export type TraceContextType = {
+  op?: string;
+  type?: 'trace';
+  span_id?: string;
+  trace_id?: string;
+  parent_span_id?: string;
+};
