+++ src/sentry/static/sentry/app/views/eventsV2/eventDetails/transaction/opsBreakdown.tsx
@@ -7,10 +7,10 @@ import {
   SentryTransactionEvent,
   SpanEntry,
   RawSpanType,
+  TraceContextType,
 } from 'app/components/events/interfaces/spans/types';
 import {SectionHeading} from 'app/components/charts/styles';
 import {pickSpanBarColour} from 'app/components/events/interfaces/spans/utils';
-import {TraceContextType} from 'app/components/events/interfaces/spans/traceView';
 import {t} from 'app/locale';
 import space from 'app/styles/space';
